requirements_base.txt
appdirs==1.4.2
asgi-redis==0.14.1
asgiref==0.14.0
asn1crypto==0.22.0
attrs==17.4.0
autobahn==0.16.0
Babel==2.3.4
boto==2.42.0
boto3==1.5.8
botocore==1.8.35
cffi==1.10.0
channels==0.17.3
colorama==0.3.7
contextlib2==0.5.4
coverage==4.2
cryptography==2.0
daphne==0.14.3
dj-database-url==0.4.1
Django==1.8.8
django-countries==4.0
django-datetime-widget==0.9.3
django-easymoney==0.7.2
django-floppyforms==1.7.0
django-idmap==1.0
django-multiselectfield==0.1.8
django-user-agents==0.3.2
django-vanilla-views==1.0.4
django-widget-tweaks==1.4.1
djangorestframework==3.4.6
docutils==0.13.1
honcho==0.7.1
huey==1.2.0
idna==2.5
IPy==0.83
jmespath==0.9.3
mock==2.0.0
msgpack-python==0.4.8
numpy==1.14.0
otree-boto2-shim==0.3.2
otree-core==1.4.40
otree-custom-export==0.1.9
otree-save-the-change==1.1.3
packaging==16.8
pbr==1.10.0
py==1.4.31
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.18
pyOpenSSL==17.2.0
pyparsing==2.1.10
pytest==2.9.2
pytest-django==3.0.0
python-dateutil==2.6.1
python-redis-lock==3.2.0
pytz==2016.6.1
PyYAML==3.12
raven==5.25.0
redis==2.10.5
requests==2.11.1
s3transfer==0.1.10
schema==0.6.2
service-identity==17.0.0
six==1.10.0
termcolor==1.1.0
Twisted==16.2.0
txaio==2.5.1
ua-parser==0.7.3
unicodecsv==0.14.1
user-agents==1.1.0
whitenoise==3.2.1
ws4py==0.3.5
XlsxWriter==0.9.3
zope.interface==4.2.0
